code cleanup
git-svn-id: http://svn.code.sf.net/p/htmlunit/code/trunk/htmlunit@15201 5f5364db-9458-4db8-a492-e30667be6df6

+++ src/main/java/com/gargoylesoftware/htmlunit/javascript/host/event/PointerEvent.java
@@ -99,7 +99,7 @@ public class PointerEvent extends MouseEvent {
                 value = String.valueOf(value);
             }
             else if (defaulValue instanceof Double) {
-                value = (double) Context.toNumber(value);
+                value = Context.toNumber(value);
             }
             else if (defaulValue instanceof Number) {
                 value = (int) Context.toNumber(value);
